Keep WordPress from corrupting post GUIDs
WordPress's default filtering corrupts post GUIDs containing ampersands
and certain other characters or substrings.  Imposer is intended to
work with fully trusted input and produce consistent results, so it
now verifies that the GUID in the DB matches the GUID it was given
to insert or update, and forcibly updates the database if necessary.

Normally, WP doesn't allow post GUIDs to be changed, but imposer
now does, as there are some scenarios where it is useful to forcibly
overwrite an existing UUID (e.g. option reference GUIDs in Postmark).